fix: align CControlEditBox with ground truth (multiple differences)
- stBlockStair: use unsigned int instead of size_t (ground truth uses
  _DWORD; size_t breaks layout on 64-bit builds)
- GetMaxTextSize: return int to match ground truth signature
- CreateChildren: set caret text color BEFORE SetText (ground truth
  writes this[909] before calling SetText)
- PrepareDrawing: when m_focus && !m_writable, return without calling
  CControlBase::PrepareDrawing (ground truth does not call base in
  this path)
- GetCaretPos: only set m_caretIndex in multi-line branch (ground truth
  sets this[35] = v7 only inside the IsMultiLine block)

+++ inc/UI/CControlEditBox.h
@@ -20,7 +20,8 @@ extern DCTIMMList g_IMMList;
 struct stBlockStair {
     // 每一行的「該行起點的位元組數」與「該行選取長度（位元組）」。
     // 反編譯碼以 pair 陣列使用，這裡保留 10 行的空間（與 GetCharByteByLine 的 10 相呼應）。
-    size_t len[10 * 2]{}; // [pos0,len0, pos1,len1, ...]
+    // 對齊反編譯：ground truth 使用 _DWORD（32-bit），非 size_t（64-bit 時佈局不同）
+    unsigned int len[10 * 2]{}; // [pos0,len0, pos1,len1, ...]
 };
 
 /**
@@ -60,7 +61,7 @@ public:
     char* GetText();
 
     // 量測
-    uint16_t GetMaxTextSize() const;
+    int GetMaxTextSize() const;  // 對齊反編譯：回傳 int（非 uint16_t）
     unsigned int GetCurTextSize() const;
     BOOL IsMultiLine() const;
